Move: external/mongo.go to lib/mongo.go Change: method called name
